# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "3.0.0",
        "title": "Blog API",
        "description": "Api that allows to create user and perform simple CRUD for a blog"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "tags": [
        {
            "name": "users",
            "description": "Operations related to users"
        },
        {
            "name": "blogs",
            "description": "Operations related to blogs"
        }
    ],
    "paths": {
        "/users/signup": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Register new user",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "signup",
                        "in": "body",
                        "description": "Signup user info",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserSignup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User info",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/users/signin": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Generate user token",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "body",
                        "description": "User login info",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserLoginInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated user token",
                        "schema": {
                            "$ref": "#/definitions/UserToken"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get the current user",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token from signin",
                        "required": true,
                        "type": "string",
                        "value": "Bearer <TOKEN_HERE>"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get the current user",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Update current user information",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "info",
                        "in": "body",
                        "description": "User information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/blogs/{id}": {
            "get": {
                "tags": [
                    "blogs"
                ],
                "summary": "Retreive a blog post by id",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token from signin",
                        "required": true,
                        "type": "string",
                        "value": "Bearer <TOKEN_HERE>"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get blog by id",
                        "schema": {
                            "schema": "object",
                            "$ref": "#/definitions/Blog"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "blogs"
                ],
                "summary": "Delete a blog post by id",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token from signin",
                        "required": true,
                        "type": "string",
                        "value": "Bearer <TOKEN_HERE>"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog info",
                        "schema": {
                            "schema": "object",
                            "$ref": "#/definitions/Blog"
                        }
                    }
                }
            }
        },
        "/blogs": {
            "get": {
                "tags": [
                    "blogs"
                ],
                "summary": "Retreive a blog post",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token from signin",
                        "required": true,
                        "type": "string",
                        "value": "Bearer <TOKEN_HERE>"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get blog posts",
                        "schema": {
                            "schema": "array",
                            "$ref": "#/definitions/BlogResult"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "blogs"
                ],
                "summary": "Create a blog post",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token from signin",
                        "required": true,
                        "type": "string",
                        "value": "Bearer <TOKEN_HERE>"
                    },
                    {
                        "name": "blog",
                        "in": "body",
                        "description": "Blog info",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BlogInfo"
                        }
                    }

                ],
                "responses": {
                    "200": {
                        "description": "Blog info",
                        "schema": {
                            "schema": "object",
                            "$ref": "#/definitions/Blog"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "UserLoginInfo": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "UserSignup": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "UserInfo": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "UserUpdate": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                }
            }
        },
        "UserToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "Blog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_dt": {
                    "type": "string"
                },
                "updated_dt": {
                    "type": "string"
                }
            }
        },
        "BlogResult": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "$ref": "#/definitions/Blog"
                }
            }
        },
        "BlogInfo": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        }
    }
}